Video 18 - FIx 1
I have missed 4 connections in Street_4Way. Basically all the markers in the middle should connect to all the other markers in the middle.  In one direciton I have forgotten to add it and cars were using a diagonal path thus driving on the incorrect lane if there were 3 x 4-Way road prefabs in a row.